Fix compilation errors: narrowing conversion and missing declaration

+++ components/samsung_ac/protocol.cpp
@@ -36,7 +36,7 @@ namespace esphome
         DecodeResult process_data(std::vector<uint8_t> &data, MessageTarget *target)
         {
             if (data.empty() || *data.begin() != 0x32)
-                return { DecodeResultType::Discard, data.empty() ? 0 : skip_data(data, 0) };
+                return { DecodeResultType::Discard, data.empty() ? uint16_t(0) : skip_data(data, 0) };
 
 
             DecodeResult  result = { DecodeResultType::Fill, 0 };
